+++ lib/templates.py
@@ -8,15 +8,25 @@ defaults_login_health_check = {
 }
 
 """Nagios template output for Cert health check"""
-cert_health_check_tmpl = Template("SSL_CERT ${status} - x509 certificate '${subject}' from '${issuer}' is valid until "
+cert_health_check_tmpl = Template("SSL_CERT(${type}) ${status} - x509 certificate '${subject}' from '${issuer}' is valid until "
                                   "${not_after} (expires in ${expiration_days} days) | 'SSL Metadata Cert'=${"
                                   "expiration_days};${warning};${critical};0;3650")
 
 defaults_cert_health_check = {
+    "type": "",
+    "status": "",
     "subject": "",
     "issuer": "",
     "not_after": "",
     "expiration_days": "",
     "warning": "",
     "critical": "",
+}
+
+"""Nagios template output for Cert health check - Type all"""
+cert_health_check_all_tmpl = Template("SSL_CERT(${type}) ${status}")
+
+defaults_cert_health_check_all = {
+    "type": "",
+    "status": ""
 }